% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acids.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{Example amino acids}
\format{
A tibble containing 20 rows and 6 columns.
}
\usage{
amino_acids
}
\description{
The chemical structures of the essential amino acids
\itemize{
\item ID - an identification number
\item NAME - the amino acid name
\item InChI - the InChI chemical structure
\item InChIKey - the InChIKey chemical structure
\item SMILES - the SMILES chemical structure
}
}
\keyword{datasets}
