% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionisation_products.R
\name{ionisationProducts}
\alias{ionisationProducts}
\title{Calculate ionisation products for a SMILES structure}
\usage{
ionisationProducts(SMILES, adduct_rules_table = adduct_rules())
}
\arguments{
\item{SMILES}{a valid SMILES string}

\item{adduct_rules_table}{table of adduct rules. Defaults to adducts()}
}
\value{
A tibble containing the mass to charge ratios of electrospray ionisation products for the specified SMILES structure.
}
\description{
Calculate electrospray ionisation product \emph{m/z} for a SMILES structure.
}
\examples{
ionisationProducts(amino_acids$SMILES[1])
}
