% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_database.R
\docType{class}
\name{MetaboliteDatabase-class}
\alias{MetaboliteDatabase-class}
\title{MetaboliteDatabase S4 class}
\description{
An S4 class for a metabolite ionisation database.
}
\section{Slots}{

\describe{
\item{\code{entries}}{database metabolite entries}

\item{\code{descriptors}}{chemical descriptors of metabolite database entries}
}}

