% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{smartsSearch}
\alias{smartsSearch}
\title{SMARTS substructure search}
\usage{
smartsSearch(SMILES, SMARTS)
}
\arguments{
\item{SMILES}{a valid SMILES structure string}

\item{SMARTS}{a valid SMARTS symbol}
}
\value{
The number of SMARTS matches.
}
\description{
Perform a SMARTS substructure searches a SMILES chemical structure.
}
\examples{
smartsSearch("C[C@H](C(=O)O)N","[OX2H]")
}
