% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{smilesToMF}
\alias{smilesToMF}
\alias{smilesToAccurateMass}
\title{Convert a SMILES structure to a molecular formula or accurate mass}
\usage{
smilesToMF(SMILES)

smilesToAccurateMass(SMILES)
}
\arguments{
\item{SMILES}{a valid SMILES structure string}
}
\description{
Convert a SMILES structure to a molecular formula or accurate mass.
}
\examples{
## Convert a SMILES structure to a molecular formula
smilesToMF("C[C@H](C(=O)O)N")

## Convert a SMILES structure to an accurate mass
smilesToAccurateMass("C[C@H](C(=O)O)N")
}
