% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{smilesToAccurateMass}
\alias{smilesToAccurateMass}
\title{Convert a SMILES structure to accurate mass}
\usage{
smilesToAccurateMass(SMILES)
}
\arguments{
\item{SMILES}{a valid SMILES structure string}
}
\description{
convert a smile to an accurate mass
}
\examples{
smilesToAccurateMass("C[C@H](C(=O)O)N")
}
