% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{smilesToMF}
\alias{smilesToMF}
\title{Convert a SMILES structure to a molecular formula}
\usage{
smilesToMF(SMILES)
}
\arguments{
\item{SMILES}{a valid SMILES structure string}
}
\description{
Convert a SMILES structure to a molecular formula.
}
\examples{
smilesToMF("C[C@H](C(=O)O)N")
}
