% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert chemical notation}
\usage{
convert(
  input,
  input_type = c("SMILES", "INCHI"),
  output_type = c("INCHI", "INCHIKEY", "SMILES")
)
}
\arguments{
\item{input}{a valid SMILE or Inchi}

\item{input_type}{either "smiles" or "inchi", denoting the input type}

\item{output_type}{either "smiles", "inchi" or "inchikey", denoting the output type}
}
\description{
convert between SMILES and Inchi and to InchiKey
}
\examples{
convert("C[C@H](C(=O)O)N",'SMILES','INCHI')
}
